# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: CWP Detailers + Kleansolar Business Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /analytics/staff-leaderboard:
    get:
      operationId: getStaffLeaderboard
      tags: [analytics]
      summary: Get staff performance leaderboard
      parameters:
        - name: month
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Staff leaderboard
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StaffLeaderboardEntry"
...
components:
  schemas:
...
    StaffLeaderboardEntry:
      type: object
      properties:
        rank:
          type: integer
        staffId:
          type: integer
        staffName:
          type: string
        jobsCompleted:
          type: integer
        revenueGenerated:
          type: number
        averageRating:
          type: number
        efficiencyScore:
          type: number
      required: [rank, staffId, staffName, jobsCompleted]
...